Add an owner role on project creation

+++ lib/screens/edit_roles/widgets/tabs/manage/manage_tab.dart
@@ -18,9 +18,10 @@ class ManageTab extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     return Column(
       children: [
-        DeleteRoleTile(
-          role: role,
-        ),
+        if (role.id != 'owner')
+          DeleteRoleTile(
+            role: role,
+          ),
         NumberOfMembersThatHaveRoleTile(
           role: role,
         ),
